src/main/python/requirements.txt
numpy>=1.21.6
Cython>=0.29.32
pycocotools==2.0.6
opencv-python>=4.5.5.64
mediapipe>=0.8.9.1
gdown>=4.4.0
yacs>=0.1.8
numba>=0.53.1
scikit-image>=0.17.2
filterpy>=1.4.5
ipython>=7.16.3
einops>=0.4.1
timm>=0.5.4
pandas>=1.3.5
zmq==0.0.0
local-attention==1.4.3                                 
scikit-learn==1.0.2
scipy==1.7.3
pytorch-lightning==1.5
mlflow==1.30
torchmetrics==0.11.1
seaborn==0.12.2
deep-sort-realtime==1.3.2
open3d==0.17.0
loguru==0.7.0
smplx==0.1.28
fvcore==0.1.5.post20221221
portalocker==2.7.0
termcolor==2.3.0
iopath==0.1.10
trimesh==3.21.5
PyOpenGL==3.1.0
freetype-py==2.3.0
pyglet==1.5.27
pyrender==0.1.45
torchvision==0.13
optuna==3.5.0